passes-ui-core: split BidiIsolationTest from ArgbColorTest (review nit)
PR #46 review flagged that ArgbColorTest.kt held two `isolated()` tests, so
a future contributor grepping for BidiIsolationTest would not find them. Split
the file along its actual responsibility lines: ArgbColorTest covers the value
class, BidiIsolationTest covers the FSI/PDI fence. Both ride the same module's
JVM unit-test source set; assertions and behaviour are unchanged.
